+++ lib/resolver/test.js
@@ -1,5 +1,6 @@
+#!/usr/bin/env node
 const { resolve, SourceType } = require("./index");
-const { deepEqual } = require("assert";)
+const { deepEqual } = require("assert");
 const snapshots = [
   {
     name: "name",
