+++ lib/resolver/tsconfig.json
@@ -1,0 +1,20 @@
+{
+  "compilerOptions": {
+    "target": "es5",
+    "module": "commonjs",
+    "lib": ["es6"],
+    // "allowJs": true,
+    // "downlevelIteration": true,
+    "preserveConstEnums": true,
+    // "alwaysStrict": true,
+    "noImplicitAny": true,
+    "noImplicitReturns": true,
+    "noImplicitThis": true,
+    // "noEmitOnError": true,
+    "strictNullChecks": true,
+    // "experimentalDecorators": true,
+    "outDir": "./dist",
+    "declaration": true
+  },
+  "include": ["./src/index.ts"]
+  }
